+++ NEWS.md
@@ -3,6 +3,7 @@
 ## April 2020
 
 * `extraOptsRT`/`extraOptsGroup` arguments for OpenMS feature grouping to allow custom command line options.
+* `importFeatureGroupsBrukerTASQ`: now correctly groups suspects across analyses if they were found >1 in the same analyses.
 
 
 ## March 2020